vaultiq/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Configuration
pydantic-settings==2.7.1
pydantic[email]==2.10.5

# Logging
structlog==24.4.0

# Document Parsing (Phase 2)
PyMuPDF==1.25.3
pdfplumber==0.11.4
python-docx==1.1.2

# Vector Store (Phase 2)
chromadb==0.6.3

# LLM (Phase 2)
ollama==0.4.7

# Benchmarking (Phase 4)
psutil==6.1.1

# Rate Limiting
slowapi==0.1.9

# File Validation
python-magic-bin==0.4.14

# Search (Phase 3)
rank-bm25==0.2.2

# JSON Schema Validation (Phase 3)
jsonschema==4.23.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
httpx==0.28.1

# HTTP
aiofiles==24.1.0
